CORE-8621 Move metadata cells into cells package
